--- docs/Komponentengalerie/managed-sample-solutions.md
@@ -1,30 +1,0 @@
----
-page_id: f632a479-cfe4-4a98-848a-a0dd0e0f6535
----
-
-import SampleSolutionTaskState, { StateDependent } from '@tdev-components/documents/SampleSolutionTaskState';
-import Solution from '@tdev-components/documents/Solution';
-import BrowserWindow from '@tdev-components/BrowserWindow';
-
-# Gesteuerte Musterlösungen
-
-<BrowserWindow>
-:::note[1. Aufgabe]
-<SampleSolutionTaskState id="df3313a5-c18f-4220-9dfe-cf4314c1b7b9" solutionId="e92b6f49-396e-48bc-8a6c-4ca94947210d">
-</SampleSolutionTaskState>
-
-Erstelle eine Lösung für die Aufgabe.
-
-<StateDependent sampleSolutionTaskStateId="df3313a5-c18f-4220-9dfe-cf4314c1b7b9">
-<Solution id="e92b6f49-396e-48bc-8a6c-4ca94947210d">
-Hallo Welt 🌍
-</Solution>
-</StateDependent>
-:::
-
-<StateDependent sampleSolutionTaskStateId="df3313a5-c18f-4220-9dfe-cf4314c1b7b9" alwaysVisibleForTeacher={false}>
-:::warning[Selbstständig korrigieren]
-Vergleiche deine Lösung nun mit der Musterlösung und korrigiere deine Antwort. Wenn die Musterlösung noch nicht freigeschaltet ist, musst du kurz warten.
-:::
-</StateDependent>
-</BrowserWindow>
