+++ docs/Komponentengalerie/self-check.md
@@ -1,0 +1,31 @@
+---
+page_id: f632a479-cfe4-4a98-848a-a0dd0e0f6535
+---
+
+import SelfCheckTaskState from '@tdev-components/documents/SelfCheck/SelfCheckTaskState';
+import SelfCheckContainer from '@tdev-components/documents/SelfCheck/SelfCheckContainer';
+import Solution from '@tdev-components/documents/Solution';
+import BrowserWindow from '@tdev-components/BrowserWindow';
+
+# Selfcheck
+
+<BrowserWindow>
+:::note[1. Aufgabe]
+<SelfCheckTaskState id="df3313a5-c18f-4220-9dfe-cf4314c1b7b9" solutionId="e92b6f49-396e-48bc-8a6c-4ca94947210d">
+</SelfCheckTaskState>
+
+Erstelle eine Lösung für die Aufgabe.
+
+<SelfCheckContainer taskStateId="df3313a5-c18f-4220-9dfe-cf4314c1b7b9">
+<Solution id="e92b6f49-396e-48bc-8a6c-4ca94947210d">
+Hallo Welt 🌍
+</Solution>
+</SelfCheckContainer>
+:::
+
+<SelfCheckContainer sampleSolutionTaskStateId="df3313a5-c18f-4220-9dfe-cf4314c1b7b9" alwaysVisibleForTeacher={false}>
+:::warning[Selbstständig korrigieren]
+Vergleiche deine Lösung nun mit der Musterlösung und korrigiere deine Antwort. Wenn die Musterlösung noch nicht freigeschaltet ist, musst du kurz warten.
+:::
+</SelfCheckContainer>
+</BrowserWindow>
